Extract arm specifc features form vulkan consumer base
In order to minimize conflicts with the codebase selected arm features are extracted to separate file.
A separate friend class encapsulating arm functionalities was created.
An object of this class is instantiated inside Vulkan Consumer Base.
This should not impact funcitionality in any way, but will reduce the impact of upstream changes on arm features.

Extracted features:
* Use frame boundary ext
* Device address replacement in fill memory/push constants
* LogFrameDebugInfo utility
* Disable subpass fusion
* Marking layers support
* Device fault support

Change-Id: I22083276093995d2f93c8396d190ac9a1ffd01ca