Fix `VulkanRebindAllocator::AllocateAHBMemory`
According to the Vulkan specification:

(VUID-VkMemoryAllocateInfo-allocationSize-02383)
If the parameters define an import operation and the external handle type
is `VK_EXTERNAL_MEMORY_HANDLE_TYPE_ANDROID_HARDWARE_BUFFER_BIT_ANDROID`,
`allocationSize` must be the size returned by
`vkGetAndroidHardwareBufferPropertiesANDROID` for the Android
hardware buffer.

As per the current state of the code, this function can only be reached
if `memory_alloc_info->ahb` is not `nullptr`, so this applies.
Thus `allocationSize` must neither be the one returned by
`vkGetImageMemoryRequirements` nor the capture time allocation size.

I also added a `GFXRECON_ASSERT` to make sure that `memory_alloc_info->ahb`
cannot be `nullptr`.

Change-Id: Iaae7e070437403c3177945cf07d7f35506ba59b6