Fix for FixDeviceAddress
Addresses a posibility of losing address-id  mapping, when several
acceleration structures reuse same acceleration structure address

In this case, this happened because of the destruction of the first one,
and then reusing of this memory to create a never one. Therefore,
we were either writing an id of something that does not exist at the moment
of issuing a FixDeviceAddress or not writing anything at all.

Fixes Arena Breakout OMM replay.

Change-Id: I4a00bc232f9a61f3ebff1696886cb8997b81c955