Make `FpsInfo` API-agnostic
As frames are defined in an API-agnostic way in GFXReconstruct,
nothing forces `FpsInfo` to be API dependent.

In particular, consumers shouldn't have access to `FpsInfo` (at
least not in it's current way) and markers should be dispatched
to `FpsInfo` through the file processor directly. Doing this
reduces code complexity and readability.

Also, measurement options are thus not Vulkan dependent and
should be enabled no matter the API

Change-Id: I271e2d9d82c1de81bfcf8afa5821429c955d4b8b