Rework tracking of AccelerationStructures (#2399)
our tracking and internal book-keeping of raytracing acceleration-structures
is based on VkAccelerationStructureKHR handles and assumes a 1:1 relation of handles and corresponding VkDeviceAddress.

this assumption can be easily violated though, e.g.:
- multiple VkAccelerationStructureKHR alias over same buffer/offset -> same AS VkDeviceAddress
 -> this works and is done in reality, but not properly covered by spec (yet)
- deleted handles, but buffer-content 'can' still be used e.g. as referenced bottom-lvl structure during top-lvl builds

so the relation VkAccelerationStructureKHR <-> VkDeviceAddress is not (1:1) but rather (0-N:1)

this is reworked in this PR:
VkAccelerationStructureKHR treated more like a lightweight 'view' on a resource.
we fully realize and use the close relationship of buffer-device-addresses and acceleration-addresses.
one can be deduced from the other. we use this when necessary and enforce with asserts.

acceleration-structure tracking is now based on VkDeviceAddress and the
lifetime of associated internal assets is now tied to the referenced VkBuffer, not the VkAccelerationStructureKHR

related issue: #2347

note:
all changes in this PR affect replay (w. -m rebind) only and fix issues with multiple and/or deleted AS-handles.
-> remaining issues during capture, when trimming AS, will be handled in a follow-up PR

Change-Id: I05b131bdebce6161402a40c2421529ecb8ca7be3