Fix virtual swapchain memory barriers
The virtual swapchain mechanism needs to copy the capture time
swapchain image (virtual image) to the replay time swpachain
imae (swapchain image).

To do this, `VulkanVirtualSwapchain` needs to do the following:
- A layout transition of the virtual image for copy (source)
- A layout transition of the swapchain image for copy (destination)
- The actual copy
- A layout transition of the virtual image back to present layout
- A layout transition of the swapchain image back to present layout

This was done correctly, but the layout transition being done using
`vkCmdPipelineBarrier`, it is important to specify the correct
flags for the memory dependency not to introduce additional
synchronization that could slow down the replay.

This copy is mainly synchronized by the calls the application has
to do at capture time in order to call `vkQueuePresentKHR`. Thus
some of the flags currently used in the `vkCmdPipelineBarrier`,
that would be valid in a general copy out of context, are actually
unnecessary and adding potential synchronization at replay time.

This is what this commit attempts to fix.

Change-Id: I5574790d6f24ed7bc9b3c9da7e71c5602f2de4e1