Implement tracking of sparse image binding
Currently there is a "gap" in the implementation of sparse
support: Tracking of the state of bindings of a sparse image.

This tracking requires intersecting the image subresource
(aspectMask, array layer, mip-map level), offset and extent
being bound with the former tracked bound state.

This is what this commit implements.

Change-Id: Ie9e32e63f9ecef0014571d3c7c41b625b7fd5780